lan8

+++ website.com/admin/product/delete.php
@@ -1,14 +1,13 @@
 <?php
-
-    $id = $_GET['id'];
-    $sql = "DELETE FROM product WHERE id = " . $id;
-    $result = mysqli_query($connect, "SELECT * FROM product WHERE id =  $id");
-    $row   = mysqli_fetch_assoc($result);
-    $old_image = ROOT_PATH . '/uploads/' .$row['image'];
-    unlink($old_image);
-    mysqli_query($connect, $sql);
-    
-    $_SESSION["message"] = "Đã xóa thành công id = $id";
-    header("Location: $base_url/admin");
+    $id = isset($_GET['id']) ? $_GET['id'] : '';
+    if($id != '') {
+        $where[] = ['id', '=', $id];
+        $db = new DB();
+        $db->delete($where);
+        $old_image = ROOT_PATH . '/uploads/' .$obj->image;
+        unlink($old_image);
+        $_SESSION["message"] = "Đã xóa thành công id = $id";
+        header("Location: $base_url/admin");
+    }
 ?>
